webusb/idlharness-expected.txt

This is a testharness.js-based test.
Found 163 tests; 2 PASS, 161 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL Navigator interface: attribute usb assert_true: The prototype object must have a property "usb" expected true got false
PASS Navigator must be primary interface of navigator 
PASS Stringification of navigator 
FAIL Navigator interface: navigator must inherit property "usb" with the proper type (0) assert_inherits: property "usb" not found in prototype chain
FAIL USB interface: existence and properties of interface object assert_own_property: self does not have own property "USB" expected property "USB" missing
FAIL USB interface object length assert_own_property: self does not have own property "USB" expected property "USB" missing
FAIL USB interface object name assert_own_property: self does not have own property "USB" expected property "USB" missing
FAIL USB interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USB" expected property "USB" missing
FAIL USB interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USB" expected property "USB" missing
FAIL USB interface: attribute onconnect assert_own_property: self does not have own property "USB" expected property "USB" missing
FAIL USB interface: attribute ondisconnect assert_own_property: self does not have own property "USB" expected property "USB" missing
FAIL USB interface: operation getDevices() assert_own_property: self does not have own property "USB" expected property "USB" missing
FAIL USB interface: operation requestDevice(USBDeviceRequestOptions) assert_own_property: self does not have own property "USB" expected property "USB" missing
FAIL USB must be primary interface of navigator.usb assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of navigator.usb assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL USB interface: navigator.usb must inherit property "onconnect" with the proper type (0) assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL USB interface: navigator.usb must inherit property "ondisconnect" with the proper type (1) assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL USB interface: navigator.usb must inherit property "getDevices" with the proper type (2) assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL USB interface: navigator.usb must inherit property "requestDevice" with the proper type (3) assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL USB interface: calling requestDevice(USBDeviceRequestOptions) on navigator.usb with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL USBConnectionEvent interface: existence and properties of interface object assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
FAIL USBConnectionEvent interface object length assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
FAIL USBConnectionEvent interface object name assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
FAIL USBConnectionEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
FAIL USBConnectionEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
FAIL USBConnectionEvent interface: attribute device assert_own_property: self does not have own property "USBConnectionEvent" expected property "USBConnectionEvent" missing
FAIL USBDevice interface: existence and properties of interface object assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface object length assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface object name assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute usbVersionMajor assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute usbVersionMinor assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute usbVersionSubminor assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute deviceClass assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute deviceSubclass assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute deviceProtocol assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute vendorId assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute productId assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute deviceVersionMajor assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute deviceVersionMinor assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute deviceVersionSubminor assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute manufacturerName assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute productName assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute serialNumber assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute configuration assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute configurations assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: attribute opened assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation open() assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation close() assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation selectConfiguration(octet) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation claimInterface(octet) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation releaseInterface(octet) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation selectAlternateInterface(octet,octet) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation controlTransferIn(USBControlTransferParameters,unsigned short) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation controlTransferOut(USBControlTransferParameters,BufferSource) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation clearHalt(USBDirection,octet) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation transferIn(octet,unsigned long) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation transferOut(octet,BufferSource) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation isochronousTransferIn(octet,[object Object]) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation isochronousTransferOut(octet,BufferSource,[object Object]) assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBDevice interface: operation reset() assert_own_property: self does not have own property "USBDevice" expected property "USBDevice" missing
FAIL USBInTransferResult interface: existence and properties of interface object assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
FAIL USBInTransferResult interface object length assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
FAIL USBInTransferResult interface object name assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
FAIL USBInTransferResult interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
FAIL USBInTransferResult interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
FAIL USBInTransferResult interface: attribute data assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
FAIL USBInTransferResult interface: attribute status assert_own_property: self does not have own property "USBInTransferResult" expected property "USBInTransferResult" missing
FAIL USBInTransferResult must be primary interface of new USBInTransferResult("ok") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBInTransferResult is not defined"
FAIL Stringification of new USBInTransferResult("ok") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBInTransferResult is not defined"
FAIL USBInTransferResult interface: new USBInTransferResult("ok") must inherit property "data" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBInTransferResult is not defined"
FAIL USBInTransferResult interface: new USBInTransferResult("ok") must inherit property "status" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBInTransferResult is not defined"
FAIL USBOutTransferResult interface: existence and properties of interface object assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
FAIL USBOutTransferResult interface object length assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
FAIL USBOutTransferResult interface object name assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
FAIL USBOutTransferResult interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
FAIL USBOutTransferResult interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
FAIL USBOutTransferResult interface: attribute bytesWritten assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
FAIL USBOutTransferResult interface: attribute status assert_own_property: self does not have own property "USBOutTransferResult" expected property "USBOutTransferResult" missing
FAIL USBOutTransferResult must be primary interface of new USBOutTransferResult("ok") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBOutTransferResult is not defined"
FAIL Stringification of new USBOutTransferResult("ok") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBOutTransferResult is not defined"
FAIL USBOutTransferResult interface: new USBOutTransferResult("ok") must inherit property "bytesWritten" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBOutTransferResult is not defined"
FAIL USBOutTransferResult interface: new USBOutTransferResult("ok") must inherit property "status" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBOutTransferResult is not defined"
FAIL USBIsochronousInTransferPacket interface: existence and properties of interface object assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
FAIL USBIsochronousInTransferPacket interface object length assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
FAIL USBIsochronousInTransferPacket interface object name assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
FAIL USBIsochronousInTransferPacket interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
FAIL USBIsochronousInTransferPacket interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
FAIL USBIsochronousInTransferPacket interface: attribute data assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
FAIL USBIsochronousInTransferPacket interface: attribute status assert_own_property: self does not have own property "USBIsochronousInTransferPacket" expected property "USBIsochronousInTransferPacket" missing
FAIL USBIsochronousInTransferPacket must be primary interface of new USBIsochronousInTransferPacket("ok") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferPacket is not defined"
FAIL Stringification of new USBIsochronousInTransferPacket("ok") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferPacket is not defined"
FAIL USBIsochronousInTransferPacket interface: new USBIsochronousInTransferPacket("ok") must inherit property "data" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferPacket is not defined"
FAIL USBIsochronousInTransferPacket interface: new USBIsochronousInTransferPacket("ok") must inherit property "status" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferPacket is not defined"
FAIL USBIsochronousInTransferResult interface: existence and properties of interface object assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
FAIL USBIsochronousInTransferResult interface object length assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
FAIL USBIsochronousInTransferResult interface object name assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
FAIL USBIsochronousInTransferResult interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
FAIL USBIsochronousInTransferResult interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
FAIL USBIsochronousInTransferResult interface: attribute data assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
FAIL USBIsochronousInTransferResult interface: attribute packets assert_own_property: self does not have own property "USBIsochronousInTransferResult" expected property "USBIsochronousInTransferResult" missing
FAIL USBIsochronousInTransferResult must be primary interface of new USBIsochronousInTransferResult([]) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferResult is not defined"
FAIL Stringification of new USBIsochronousInTransferResult([]) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferResult is not defined"
FAIL USBIsochronousInTransferResult interface: new USBIsochronousInTransferResult([]) must inherit property "data" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferResult is not defined"
FAIL USBIsochronousInTransferResult interface: new USBIsochronousInTransferResult([]) must inherit property "packets" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousInTransferResult is not defined"
FAIL USBIsochronousOutTransferPacket interface: existence and properties of interface object assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
FAIL USBIsochronousOutTransferPacket interface object length assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
FAIL USBIsochronousOutTransferPacket interface object name assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
FAIL USBIsochronousOutTransferPacket interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
FAIL USBIsochronousOutTransferPacket interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
FAIL USBIsochronousOutTransferPacket interface: attribute bytesWritten assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
FAIL USBIsochronousOutTransferPacket interface: attribute status assert_own_property: self does not have own property "USBIsochronousOutTransferPacket" expected property "USBIsochronousOutTransferPacket" missing
FAIL USBIsochronousOutTransferPacket must be primary interface of new USBIsochronousOutTransferPacket("ok") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferPacket is not defined"
FAIL Stringification of new USBIsochronousOutTransferPacket("ok") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferPacket is not defined"
FAIL USBIsochronousOutTransferPacket interface: new USBIsochronousOutTransferPacket("ok") must inherit property "bytesWritten" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferPacket is not defined"
FAIL USBIsochronousOutTransferPacket interface: new USBIsochronousOutTransferPacket("ok") must inherit property "status" with the proper type (1) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferPacket is not defined"
FAIL USBIsochronousOutTransferResult interface: existence and properties of interface object assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
FAIL USBIsochronousOutTransferResult interface object length assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
FAIL USBIsochronousOutTransferResult interface object name assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
FAIL USBIsochronousOutTransferResult interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
FAIL USBIsochronousOutTransferResult interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
FAIL USBIsochronousOutTransferResult interface: attribute packets assert_own_property: self does not have own property "USBIsochronousOutTransferResult" expected property "USBIsochronousOutTransferResult" missing
FAIL USBIsochronousOutTransferResult must be primary interface of new USBIsochronousOutTransferResult([]) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferResult is not defined"
FAIL Stringification of new USBIsochronousOutTransferResult([]) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferResult is not defined"
FAIL USBIsochronousOutTransferResult interface: new USBIsochronousOutTransferResult([]) must inherit property "packets" with the proper type (0) assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: USBIsochronousOutTransferResult is not defined"
FAIL USBConfiguration interface: existence and properties of interface object assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
FAIL USBConfiguration interface object length assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
FAIL USBConfiguration interface object name assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
FAIL USBConfiguration interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
FAIL USBConfiguration interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
FAIL USBConfiguration interface: attribute configurationValue assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
FAIL USBConfiguration interface: attribute configurationName assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
FAIL USBConfiguration interface: attribute interfaces assert_own_property: self does not have own property "USBConfiguration" expected property "USBConfiguration" missing
FAIL USBInterface interface: existence and properties of interface object assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
FAIL USBInterface interface object length assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
FAIL USBInterface interface object name assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
FAIL USBInterface interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
FAIL USBInterface interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
FAIL USBInterface interface: attribute interfaceNumber assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
FAIL USBInterface interface: attribute alternate assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
FAIL USBInterface interface: attribute alternates assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
FAIL USBInterface interface: attribute claimed assert_own_property: self does not have own property "USBInterface" expected property "USBInterface" missing
FAIL USBAlternateInterface interface: existence and properties of interface object assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBAlternateInterface interface object length assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBAlternateInterface interface object name assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBAlternateInterface interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBAlternateInterface interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBAlternateInterface interface: attribute alternateSetting assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBAlternateInterface interface: attribute interfaceClass assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBAlternateInterface interface: attribute interfaceSubclass assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBAlternateInterface interface: attribute interfaceProtocol assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBAlternateInterface interface: attribute interfaceName assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBAlternateInterface interface: attribute endpoints assert_own_property: self does not have own property "USBAlternateInterface" expected property "USBAlternateInterface" missing
FAIL USBEndpoint interface: existence and properties of interface object assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
FAIL USBEndpoint interface object length assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
FAIL USBEndpoint interface object name assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
FAIL USBEndpoint interface: existence and properties of interface prototype object assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
FAIL USBEndpoint interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
FAIL USBEndpoint interface: attribute endpointNumber assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
FAIL USBEndpoint interface: attribute direction assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
FAIL USBEndpoint interface: attribute type assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
FAIL USBEndpoint interface: attribute packetSize assert_own_property: self does not have own property "USBEndpoint" expected property "USBEndpoint" missing
Harness: the test ran to completion.
